added makePedReload server side - Makes a ped or player reload Syntax: makePedReload ( ped thePed ) Reloads the current weapon.
Note: this will fail but return true if
1) the ped is crouched and moving
2) the ped is using a weapon without clip ammo (or minigun/flamethrower/fire extinguisher)
3) the ped is using his weapon (shooting/aiming)
4) the ped moved while crouching recently

The reason for this is that if the above where not disabled it would cancel the animation and your player would be able to reload instantly.
For the same reason while crouched and reloading you will not be able to move unless you stand up.