Added getPedMoveState clientside: string getPedMoveState ( ped thePed )
Returns:
"stand" - Ped is standing still
"walk" - Ped is walking
"powerwalk" - Ped is walking quickly
"jog" - Ped is jogging
"sprint" - Ped is sprinting
"crouch" - Ped is crouching still.
'false' if the ped is not streamed in, or the movement type is unknown, or the ped is in a vehicle.

In future we should look into adding movement types for crouching, since moving while crouch will simply return false.