Added onDebugMessage and onClientDebugMessage
These events get triggered when an error, warning, info or custom debug message should appear when "debugscript" command was executed. However, these events work even if this command was never used. This allows scripters to make their own debugging tools, for instance for remote viewing of errors and warnings, or to simply make a neater and/or more controllable debug output window.

Parameters: string message, int level, string file, int line

The parameters should be pretty straightforward.

Please note that "file" and "line" may return nil, when the source could not be found. (For instance in the "Event triggered, but not added" errors)